refactor: correct update data functions
